Refactor Frontend into modular components with improved separation of concerns
